services/variant/requirements.txt
alembic==1.4.2
grpcio==1.30.0
grpcio-tools==1.30.0
psycopg2-binary==2.8.5
SQLAlchemy==1.3.18